Create C-Tsundoku.cpp
We have two desks: A and B. Desk A has a vertical stack of
N books on it, and Desk B similarly has
M books on it.

It takes us
A
i
​
  minutes to read the
i-th book from the top on Desk A
(1≤i≤N), and
B
i
​
  minutes to read the
i-th book from the top on Desk B
(1≤i≤M).

Consider the following action:

Choose a desk with a book remaining, read the topmost book on that desk, and remove it from the desk.
How many books can we read at most by repeating this action so that it takes us at most
K minutes in total? We ignore the time it takes to do anything other than reading.